[skip ci] Refactor post_data.js and format_data.js into modules specific to each chat provider
